SEARCH-444
- Created an api to clear the user index data folder
