+++ js/preload/preloadMain.js
@@ -137,6 +137,11 @@ function createAPI() {
         Search: remote.require('./search/search.js').Search,
 
         /**
+         * Function to clear the user index data
+         */
+        clearIndexFolder: remote.require('./search/search.js').clearIndexFolder,
+
+        /**
          * Brings window forward and gives focus.
          * @param  {String} windowName Name of window. Note: main window name is 'main'
          */
